requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncio-pool==0.6.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
frozenlist==1.7.0
geomet==1.1.0
greenlet==3.2.3
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.3
openai==1.97.1
passlib==1.7.4
propcache==0.3.2
psycopg-binary==3.2.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.12.0a1
pydantic_core==2.37.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
scylla-driver==3.29.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
